requirements.txt
pyyaml>=6.0
scapy>=2.5.0
python-nmap>=0.7.1
requests>=2.31.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-prometheus>=0.40b0
prometheus-client
eero
schedule
flask
